requirements.txt
yfinance
pandas
python-telegram-bot==20.7